requirements.txt
Flask
pymongo
Werkzeug
pandas
joblib
groq
python-dotenv
flask_socketio
flask-mail
apscheduler 
pytz
